Update to fix texture issues
Evidently, texture names for "workbenches" a.k.a. buildings with interactions spots, are named to indicate the direction the *pawn* faces when interacting with it instead of the building's orientation itself. Contrary to what the official archive of textures indicates, and how almost every building and item in game is displayed. Which, frankly, is just fucking stupid. So this update reverses all the mod texture names to make them appear correctly in game. It feels like a hackish work around and I hate it.