Balancing for all three atomizers
The standard and small atomizers felt underwhelming. The goal of this mod is to make each option viable for different colony sizes and needs, not just "this one is bigger, so it's better". These subtle changes will hopefully let the small atomizer shine a little more while keeping the larger variants still very much viable.